add coordinator admin view with cs106 access
if they are the cs198 coordinator, then they should get normal access to all of the
cs106 classes. show them a special admin page which links them to the unknown sl
for each of the cs106 classes. the reason we link to the unknown sl because that is
the default that shows all students and prompts them with a view to change section leaders